+++ examples/shopping-list-tutorial/lesson-3/test/functional/index.test.js
@@ -12,12 +12,6 @@ fixture`Shopping List`.beforeEach(async (t) => {
   await t.navigateTo(MAIN_PAGE)
 })
 
-test('home page', async (t) => {
-  await t
-    .expect(await Selector('h1').withText('Shopping List').exists)
-    .eql(true)
-})
-
 test('createShoppingList', async () => {
   const command = {
     aggregateName: 'ShoppingList',
@@ -145,7 +139,7 @@ test('createShoppingItems', async () => {
   }
 })
 
-test('validation should works correctly', async () => {
+test('validation should work correctly', async () => {
   const matches = [
     {
       command: {
@@ -154,7 +148,7 @@ test('validation should works correctly', async () => {
         type: 'createShoppingList',
         payload: {},
       },
-      error: 'name is required',
+      error: 'The "name" field is required',
     },
     {
       command: {
@@ -165,7 +159,7 @@ test('validation should works correctly', async () => {
           name: 'List 1',
         },
       },
-      error: 'shopping list already exists',
+      error: 'Shopping list already exists',
     },
     {
       command: {
@@ -177,7 +171,7 @@ test('validation should works correctly', async () => {
           text: 'Bread',
         },
       },
-      error: 'shopping list does not exist',
+      error: 'Shopping list does not exist',
     },
   ]
 
@@ -196,43 +190,24 @@ test('validation should works correctly', async () => {
   }
 })
 
-test('query should work correctly', async () => {
-  const response = await fetch(
-    `${MAIN_PAGE}/api/query/shoppingList/shopping-list-1?origin=${MAIN_PAGE}`,
-    {
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      method: 'GET',
-    }
-  )
+test('read model query should work correctly', async () => {
+  const response = await fetch(`${MAIN_PAGE}/api/query/ShoppingLists/all`, {
+    headers: {
+      'Content-Type': 'application/json',
+    },
+    method: 'GET',
+  })
 
   const result = await response.json()
 
-  expect(result.data).to.deep.equal({
+  expect(result.data).to.have.lengthOf(1)
+  expect(result.data[0]).to.include({
     id: 'shopping-list-1',
     name: 'List 1',
-    list: [
-      {
-        id: '1',
-        text: 'Milk',
-        checked: false,
-      },
-      {
-        id: '2',
-        text: 'Eggs',
-        checked: false,
-      },
-      {
-        id: '3',
-        text: 'Canned beans',
-        checked: false,
-      },
-      {
-        id: '4',
-        text: 'Paper towels',
-        checked: false,
-      },
-    ],
   })
+})
+
+test('shopping list is displayed on page', async (t) => {
+  await t.expect(Selector('td').withText('1').exists).eql(true)
+  await t.expect(Selector('td').withText('List 1').exists).eql(true)
 })