+++ examples/shopping-list-tutorial/lesson-4/client/components/App.js
@@ -1,0 +1,17 @@
+import React from 'react'
+import { renderRoutes } from 'react-router-config'
+import Header from './Header'
+
+const App = ({ route, children }) => (
+  <div>
+    <Header
+      title="ReSolve Shopping List Example"
+      name="Shopping List"
+      css={['/bootstrap.min.css']}
+    />
+    {renderRoutes(route.routes)}
+    {children}
+  </div>
+)
+
+export default App