+++ examples/shopping-list-tutorial/lesson-5/config.dev.js
@@ -8,6 +8,26 @@ const devConfig = {
   staticPath: 'static',
   staticDir: 'static',
   distDir: 'dist',
+  readModelConnectors: {
+    default: {
+      module: '@resolve-js/readmodel-lite',
+      options: {
+        databaseFile: 'data/read-models.db',
+      },
+    },
+    /*
+      default: {
+        module: '@resolve-js/readmodel-mysql',
+        options: {
+          host: 'localhost',
+          port: 3306,
+          user: 'customUser',
+          password: 'customPassword',
+          database: 'customDatabaseName'
+        }
+      }
+    */
+  },
   eventstoreAdapter: {
     module: '@resolve-js/eventstore-lite',
     options: {
@@ -16,7 +36,21 @@ const devConfig = {
       snapshotBucketSize: 100,
     },
   },
-  jwtCookie: {
+  /*
+    {
+      module: '@resolve-js/eventstore-mysql',
+      options: {
+        host: 'localhost',
+        port: 3306,
+        user: 'customUser',
+        password: 'customPassword',
+        database: 'customDatabaseName',
+        eventsTableName: 'customTableName',
+        secretsDatabase: 'customSecretsDatabaseName',
+        secretsTableName: 'customSecretsTableName'
+      }
+    }
+  */ jwtCookie: {
     name: 'jwt',
     maxAge: 31536000000,
   },