+++ packages/modules/module-admin/src/commands/sagas/resume.js
@@ -13,6 +13,6 @@ export const command = 'resume <saga>'
 export const describe = 'resume saga updates'
 export const builder = (yargs) =>
   yargs.positional('saga', {
-    describe: 'an existing saga`s name',
+    describe: "an existing saga's name",
     type: 'string',
   })