+++ packages/runtime/adapters/eventstore-adapters/eventstore-base/src/export-events.ts
@@ -70,6 +70,10 @@ async function* generator(context: any): AsyncGenerator<Buffer, void> {
       context.cursor = getNextCursor(context.cursor, [event])
     }
     if (events.length === 0) {
+      context.isEnd = true
+      await endProcessEvents(context)
+      return
+    } else if (context.externalTimeout) {
       await endProcessEvents(context)
       return
     }
@@ -96,9 +100,14 @@ const exportEventsStream = <ConnectedProps extends AdapterPoolConnectedProps>(
     cursor,
     bufferSize,
     isBufferOverflow: false,
+    isEnd: false,
+    externalTimeout: false,
   }
 
   const stream: Readable = Readable.from(generator(context))
+  stream.on('timeout', () => {
+    context.externalTimeout = true
+  })
   Object.defineProperty(stream, 'cursor', {
     get() {
       return context.cursor
@@ -107,6 +116,11 @@ const exportEventsStream = <ConnectedProps extends AdapterPoolConnectedProps>(
   Object.defineProperty(stream, 'isBufferOverflow', {
     get() {
       return context.isBufferOverflow
+    },
+  })
+  Object.defineProperty(stream, 'isEnd', {
+    get() {
+      return context.isEnd
     },
   })