+++ CHANGELOG.md
@@ -23,6 +23,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - When querying read model resolvers from testing tools, the resolver results are returned unwrapped, the **data** property was removed.
 
+#### Local config
+
+- Removed the **eventBroker** section
+
 #### Cloud config
 
 - Runtime env RESOLVE_ES_SECRET_ARN renamed to RESOLVE_USER_SECRET_ARN
@@ -32,6 +36,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Runtime env RESOLVE_ES_SECRET_ARN renamed to RESOLVE_USER_SECRET_ARN
 - Runtime env RESOLVE_ES_EVENTS_TABLE removed
 - Runtime env RESOLVE_ES_SECRETS_TABLE removed
+
+### Added
+
+#### Local config
+
+- Added the **name** option. By default, it shares its value with the "name" field from package.json.
 
 ### Changed